Implemented the input system package
Added the input system package to the project.
Also fixed up the puff script for movement to work with it.
There is an error when trying to use the jump button saying that the method is not found, but the event still triggers regardless. This will need to be debugged later.